+++ api/src/services/setup/view-indexer-progress.js
